requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.6
python-dotenv==1.2.2
numpy==2.4.2